added pulse initiation code
imports image frames as jpg.  We feed it jellyfish centroid locations.  Then it finds the edge of the jellyfish, overlays radi at definable angle increments, and measures the radi each frame.  Then it groups radi in 45 degree rollups, locates the region that shortens past a threshold length first, and determines that to be the "lead" region, and ranks all the regions in order of which are shortening first.